+++ CHANGELOG.md
@@ -370,9 +370,8 @@
 * `modify` and `replaces` methods of `State` effect returns result of an applied function
 * Change precedence of `->>` method of `Traversable` - from 3 to 5
 * Rename `Stack` datastructure to `List`
-* Define `Stack` interface with `push` and `pop` methods
 * Define new type synonymous: `:=:=>`
-* Define experimental `Insert` verb for `Morphable` ability
+* Define experimental `Insert` and `Push` verbs for `Morphable` ability
 * Remove `Insertable` ability for datastructures
 * Change order of arguments in `<:.:>`, `>:.:>`, `<:.:<`, `>:.:<` type synonyms
 * Define fixity and precedence for `TU`, `UT`, `TUT` and `T_U` type synonyms